Add threaded analyzer mode, sharing data and plugins
Add worker threads to handle packets. The packets are dispatched to
threads at layer 3, based on a toepliz symmetric hash.
Since IP src/dst packets are always sent to the same workers, there is
no need to lock the flows or defrag engines. This means the
implementation is mostly lock-free.
Each worker will run all plugins (so the plugins are currently locked
using a mutex).