libpcap-analyzer: move all analyzer data to Analyzer, get rid of TAD
Move all analyzer-related data to the Analyzer struct, and update all
functions to declare a mutable reference on Analyzer.

Using mutable arguments (and since all functions are called from single
threads), there is no need for locks, cells, or thread local storage
anymore: the code is now clearer and faster.

If mutable arguments become a problem later (for ex, this prevents a
plugin from querying flows at the moment), this may be changed for a
container with inner mutability.